[#496] Refactor data unavailable field as a component
