+++ frontend/src/akvo-react-form/components/index.js
@@ -1,2 +1,3 @@
 export { default as SavedSubmissionList } from "./SavedSubmissionList";
 export { default as QuestionGroup } from "./QuestionGroup";
+export { default as DataUnavailableField } from "./DataUnavailableField";
